server/pyproject.toml
[project]
name = "dakora-server"
version = "2.0.0"
description = "Dakora Platform Server - AI Control Plane for Prompt Management"
readme = "README.md"
authors = [
    { name = "bogdanpi", email = "bogdan@iduntechnologies.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "watchdog>=6.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dakora_server"]